+++ routes/userRoute.js
@@ -1,6 +1,6 @@
 import express from "express"
-import UserController from "../controllers/user.js"
-import AuthMiddleware from "../middlewares/auth.js"
+import UserController from "../controllers/userController.js"
+import AuthMiddleware from "../middlewares/authMiddleware.js"
 
 const UserRouter = express.Router()
