+++ services/userService.js
@@ -1,4 +1,4 @@
-import User from "../models/user.js"
+import User from "../models/userModel.js"
 import mongoose from "mongoose"
 
 class UserService {
